requirements.txt
ddt==1.4.2
Django==3.2.4
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
influxdb-client==1.18.0
psycopg2-binary==2.8.6
